requirements.txt
tweepy
python-dotenv
emoji
schedule